Mybatis

+++ src/main/java/edu/eci/cvds/sampleprj/dao/ClienteDAO.java
@@ -6,8 +6,8 @@ import edu.eci.cvds.samples.entities.Cliente;
 
 public interface ClienteDAO {
 
-   public void save(int id ,int idc, int idit, Date fechainicio,Date fechafin) throws PersistenceException;
+   public void save(int id ,long idc, int idit, Date fechainicio,Date fechafin) throws PersistenceException;
 
-   public Cliente load(int id) throws PersistenceException;
+   public Cliente load(long id) throws PersistenceException;
 
 }
